refactor(agent): satisfy source complexity budgets
