+++ package.json
@@ -153,7 +153,7 @@
     "test:cli-surface-contract": "node scripts/require-shifu.mjs test:cli-surface-contract && uv run --project framework/core --frozen ruff format --check framework/core/src/python/kungfu/cli/surface_contract.py framework/core/src/python/kungfu/cli/catalog_projection.py framework/core/src/python/kungfu/cli/help_projection.py framework/core/src/python/kungfu/cli/help_manifest.py framework/core/src/python/kungfu/cli/commands/__init__.py framework/core/src/python/kungfu/cli/commands/contract.py framework/core/src/python/kungfu/cli/commands/primitive.py framework/core/src/python/kungfu/cli/commands/dev.py framework/core/tests/python/test_cli_surface_contract.py framework/core/tests/python/test_cli_progressive_help.py framework/core/tests/python/test_cli_help_manifest.py framework/core/tests/python/test_primitive_commands.py && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_cli_surface_contract.py framework/core/tests/python/test_cli_progressive_help.py framework/core/tests/python/test_cli_help_manifest.py framework/core/tests/python/test_primitive_commands.py",
     "check:cli-catalog-parity": "node scripts/require-shifu.mjs check:cli-catalog-parity && node scripts/check-cli-catalog-parity.mjs && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen python -m kungfu.cli.catalog_projection --check",
     "cli-surface:qualify": "node scripts/require-shifu.mjs cli-surface:qualify && node product/scripts/cli-surface-qualification.mjs",
-    "test:cli-surface-qualification": "node scripts/require-shifu.mjs test:cli-surface-qualification && node --test product/scripts/archive.test.mjs product/scripts/cli-launcher.test.mjs product/scripts/cli-surface-qualification.test.mjs product/scripts/dist.test.mjs",
+    "test:cli-surface-qualification": "node scripts/require-shifu.mjs test:cli-surface-qualification && node --test product/scripts/archive.test.mjs product/scripts/cli-launcher.test.mjs product/scripts/cli-surface-qualification.test.mjs product/scripts/dist.test.mjs product/scripts/dist-cli-executable-layout.test.mjs",
     "qualify:product-cache-home": "node scripts/require-shifu.mjs qualify:product-cache-home && node product/scripts/qualify-product-cache-home.mjs --json",
     "test:product-cache-home": "node scripts/require-shifu.mjs test:product-cache-home && node --test product/scripts/qualify-product-cache-home.test.mjs",
     "check:health-diagnostics-contract": "node scripts/require-shifu.mjs check:health-diagnostics-contract && node --test scripts/check-health-diagnostics-contract.test.mjs",
