+++ scripts/source-acceptance.mjs
@@ -685,6 +685,7 @@ export function sourceAcceptancePlan(files, evidenceBaseCommit = '') {
         'scripts/source-acceptance.test.mjs',
         'scripts/platform-command.test.mjs',
         'product/scripts/dist.test.mjs',
+        'product/scripts/dist-cli-executable-layout.test.mjs',
         'scripts/opencode-local-model-canary-workflow.test.mjs',
         'scripts/kungfu-workflow-authority.test.mjs',
         'scripts/code-complexity-budget.test.mjs',
